+++ km/com/hifiremote/jp1/SonyComboImporter.java
@@ -15,7 +15,7 @@ public class SonyComboImporter
     if ( parmToSet < 0 )
       return;
 
-    byte[] hex = hexData.getData();
+    int[] hex = hexData.getData();
 
     int device = 0;
     int protocol = 0;
