+++ km/com/hifiremote/jp1/XorCheck.java
@@ -44,7 +44,7 @@ public class XorCheck
 
   public void in( Value[] parms, Hex hexData, DeviceParameter[] devParms, int onlyIndex )
   {
-    byte[] hex = hexData.getData();
+    int[] hex = hexData.getData();
     // System.err.println("XorCheck(" + bits +","+ destOffset +","+ seed +","+ count +","+ sourceOffset +","+ step +").in(" + hex.length +")");
     int v = seed;
     int s = sourceOffset;
